+++ src/core/data.jl
@@ -1,4 +1,4 @@
-# tools for working with PowerModels internal data format
+# tools for working with PowerModels internal data dict structure
 
 ""
 function calc_voltage_product_bounds(buspairs)
@@ -639,6 +639,370 @@ function _check_cost_functions(id, comp)
             end
         else
             warn(LOGGER, "Unknown generator cost model of type $(comp["model"])")
+        end
+    end
+end
+
+
+
+"""
+finds active network buses and branches that are not necessary for the
+computation and sets their status to off.
+
+Works on a PowerModels data dict, so that a it can be used without a GenericPowerModel object
+
+Warning: this implementation has quadratic complexity, in the worst case
+"""
+function propagate_topology_status(data::Dict{String,Any})
+    if data["multinetwork"]
+        for (i,nw_data) in data["nw"]
+            _propagate_topology_status(nw_data)
+        end
+    else
+         _propagate_topology_status(data)
+    end
+end
+
+function _propagate_topology_status(data::Dict{String,Any})
+    buses = Dict(bus["bus_i"] => bus for (i,bus) in data["bus"])
+
+    # compute what active components are incident to each bus
+    incident_load = bus_load_lookup(data["load"], data["bus"])
+    incident_active_load = Dict()
+    for (i, load_list) in incident_load
+        incident_active_load[i] = filter(load -> load["status"], load_list)
+    end
+
+    incident_shunt = bus_shunt_lookup(data["shunt"], data["bus"])
+    incident_active_shunt = Dict()
+    for (i, shunt_list) in incident_shunt
+        incident_active_shunt[i] = filter(shunt -> shunt["status"], shunt_list)
+    end
+
+    incident_gen = bus_gen_lookup(data["gen"], data["bus"])
+    incident_active_gen = Dict()
+    for (i, gen_list) in incident_gen
+        incident_active_gen[i] = filter(gen -> gen["gen_status"] != 0, gen_list)
+    end
+
+    incident_branch = Dict(bus["bus_i"] => [] for (i,bus) in data["bus"])
+    for (i,branch) in data["branch"]
+        push!(incident_branch[branch["f_bus"]], branch)
+        push!(incident_branch[branch["t_bus"]], branch)
+    end
+
+    incident_dcline = Dict(bus["bus_i"] => [] for (i,bus) in data["bus"])
+    for (i,dcline) in data["dcline"]
+        push!(incident_dcline[dcline["f_bus"]], dcline)
+        push!(incident_dcline[dcline["t_bus"]], dcline)
+    end
+
+    updated = true
+    iteration = 0
+
+    while updated
+        while updated
+            iteration += 1
+            updated = false
+
+            for (i,branch) in data["branch"]
+                if branch["br_status"] != 0
+                    f_bus = buses[branch["f_bus"]]
+                    t_bus = buses[branch["t_bus"]]
+
+                    if f_bus["bus_type"] == 4 || t_bus["bus_type"] == 4
+                        info(LOGGER, "deactivating branch $(i):($(branch["f_bus"]),$(branch["t_bus"])) due to connecting bus status")
+                        branch["br_status"] = 0
+                        updated = true
+                    end
+                end
+            end
+
+            for (i,dcline) in data["dcline"]
+                if dcline["br_status"] != 0
+                    f_bus = buses[dcline["f_bus"]]
+                    t_bus = buses[dcline["t_bus"]]
+
+                    if f_bus["bus_type"] == 4 || t_bus["bus_type"] == 4
+                        info(LOGGER, "deactivating dcline $(i):($(dcline["f_bus"]),$(dcline["t_bus"])) due to connecting bus status")
+                        dcline["br_status"] = 0
+                        updated = true
+                    end
+                end
+            end
+
+            for (i,bus) in buses
+                if bus["bus_type"] != 4
+                    if length(incident_branch[i]) + length(incident_dcline[i]) > 0
+                        incident_branch_count = sum([0; [branch["br_status"] for branch in incident_branch[i]]])
+                        incident_dcline_count = sum([0; [dcline["br_status"] for dcline in incident_dcline[i]]])
+                        incident_active_edge = incident_branch_count + incident_dcline_count
+                    else
+                        incident_active_edge = 0
+                    end
+
+                    #println("bus $(i) active branch $(incident_active_edge)")
+                    #println("bus $(i) active gen $(incident_active_gen)")
+                    #println("bus $(i) active load $(incident_active_load)")
+                    #println("bus $(i) active shunt $(incident_active_shunt)")
+
+                    if incident_active_edge == 1 && length(incident_active_gen[i]) == 0 && length(incident_active_load[i]) == 0 && length(incident_active_shunt[i]) == 0
+                        info(LOGGER, "deactivating bus $(i) due to dangling bus without generation and load")
+                        bus["bus_type"] = 4
+                        updated = true
+                    end
+
+                else # bus type == 4
+                    for load in incident_active_load[i]
+                        if load["status"]
+                            info(LOGGER, "deactivating load $(load["index"]) due to inactive bus $(i)")
+                            load["status"] = false
+                            updated = true
+                        end
+                    end
+
+                    for shunt in incident_active_shunt[i]
+                        if shunt["status"]
+                            info(LOGGER, "deactivating shunt $(shunt["index"]) due to inactive bus $(i)")
+                            shunt["status"] = false
+                            updated = true
+                        end
+                    end
+
+                    for gen in incident_active_gen[i]
+                        if gen["gen_status"] != 0
+                            info(LOGGER, "deactivating generator $(gen["index"]) due to inactive bus $(i)")
+                            gen["gen_status"] = 0
+                            updated = true
+                        end
+                    end
+                end
+            end
+        end
+
+        ccs = connected_components(data)
+
+        #println(ccs)
+        #TODO set reference node for each cc
+
+        for cc in ccs
+            cc_active_loads = [0]
+            cc_active_shunts = [0]
+            cc_active_gens = [0]
+            
+            for i in cc
+                cc_active_loads = push!(cc_active_loads, length(incident_active_load[i]))
+                cc_active_shunts = push!(cc_active_shunts, length(incident_active_shunt[i])) 
+                cc_active_gens = push!(cc_active_gens, length(incident_active_gen[i]))
+            end
+
+            active_load_count = sum(cc_active_loads)
+            active_shunt_count = sum(cc_active_shunts)
+            active_gen_count = sum(cc_active_gens)
+
+            if (active_load_count == 0 && active_shunt_count == 0) || active_gen_count == 0
+                info(LOGGER, "deactivating connected component $(cc) due to isolation without generation and load")
+                for i in cc
+                    buses[i]["bus_type"] = 4
+                end
+                updated = true
+            end
+        end
+
+    end
+
+    info(LOGGER, "topology status propagation fixpoint reached in $(iteration) rounds")
+
+    check_refrence_buses(data)
+end
+
+
+"""
+determines the largest connected component of the network and turns everything else off
+"""
+function select_largest_component(data::Dict{String,Any})
+    if data["multinetwork"]
+        for (i,nw_data) in data["nw"]
+            _select_largest_component(nw_data)
+        end
+    else
+         _select_largest_component(data)
+    end
+end
+
+function _select_largest_component(data::Dict{String,Any})
+    ccs = connected_components(data)
+    info(LOGGER, "found $(length(ccs)) components")
+
+    ccs_order = sort(collect(ccs); by=length)
+    largest_cc = ccs_order[end]
+
+    info(LOGGER, "largest component has $(length(largest_cc)) buses")
+
+    for (i,bus) in data["bus"]
+        if bus["bus_type"] != 4 && !(bus["index"] in largest_cc)
+            bus["bus_type"] = 4
+            info(LOGGER, "deactivating bus $(i) due to small connected component")
+        end
+    end
+
+    check_refrence_buses(data)
+end
+
+
+"""
+checks that each connected components has a reference bus, if not, adds one
+"""
+function check_refrence_buses(data::Dict{String,Any})
+    if haskey(data, "multinetwork") && data["multinetwork"]
+        for (i,nw_data) in data["nw"]
+            _check_refrence_buses(nw_data)
+        end
+    else
+        _check_refrence_buses(data)
+    end
+end
+
+
+function _check_refrence_buses(data::Dict{String,Any})
+    bus_lookup = Dict(bus["bus_i"] => bus for (i,bus) in data["bus"])
+    bus_gen = bus_gen_lookup(data["gen"], data["bus"])
+
+    ccs = connected_components(data)
+    ccs_order = sort(collect(ccs); by=length)
+
+    bus_to_cc = Dict()
+    for (i, cc) in enumerate(ccs_order)
+        for bus_i in cc
+            bus_to_cc[bus_i] = i
+        end
+    end
+
+    cc_gens = Dict( i => Dict() for (i, cc) in enumerate(ccs_order) )
+    for (i, gen) in data["gen"]
+        bus_id = gen["gen_bus"]
+        if haskey(bus_to_cc, bus_id)
+            cc_id = bus_to_cc[bus_id]
+            cc_gens[cc_id][i] = gen
+        end
+    end
+
+    for (i, cc) in enumerate(ccs_order)
+        check_component_refrence_bus(cc, bus_lookup, cc_gens[i])
+    end
+end
+
+
+"""
+checks that a connected component has a reference bus, if not, adds one
+"""
+function check_component_refrence_bus(component_bus_ids, bus_lookup, component_gens)
+    refrence_buses = Set()
+    for bus_id in component_bus_ids
+        bus = bus_lookup[bus_id]
+        if bus["bus_type"] == 3
+            push!(refrence_buses, bus_id)
+        end
+    end
+
+    if length(refrence_buses) == 0
+        warn(LOGGER, "no reference bus found in connected component $(component_bus_ids)")
+
+        if length(component_gens) > 0
+            big_gen = biggest_generator(component_gens)
+            gen_bus = bus_lookup[big_gen["gen_bus"]]
+            gen_bus["bus_type"] = 3
+            warn(LOGGER, "setting bus $(gen_bus["index"]) as reference bus in connected component $(component_bus_ids), based on generator $(big_gen["index"])")
+        else
+            warn(LOGGER, "no generators found in connected component $(component_bus_ids), try running propagate_topology_status")
+        end
+    end
+end
+
+
+"builds a lookup list of what generators are connected to a given bus"
+function bus_gen_lookup(gen_data::Dict{String,Any}, bus_data::Dict{String,Any})
+    bus_gen = Dict(bus["bus_i"] => [] for (i,bus) in bus_data)
+    for (i,gen) in gen_data
+        push!(bus_gen[gen["gen_bus"]], gen)
+    end
+    return bus_gen
+end
+
+"builds a lookup list of what loads are connected to a given bus"
+function bus_load_lookup(load_data::Dict{String,Any}, bus_data::Dict{String,Any})
+    bus_load = Dict(bus["bus_i"] => [] for (i,bus) in bus_data)
+    for (i,load) in load_data
+        push!(bus_load[load["load_bus"]], load)
+    end
+    return bus_load
+end
+
+"builds a lookup list of what shunts are connected to a given bus"
+function bus_shunt_lookup(shunt_data::Dict{String,Any}, bus_data::Dict{String,Any})
+    bus_shunt = Dict(bus["bus_i"] => [] for (i,bus) in bus_data)
+    for (i,shunt) in shunt_data
+        push!(bus_shunt[shunt["shunt_bus"]], shunt)
+    end
+    return bus_shunt
+end
+
+
+"""
+computes the connected components of the network graph
+returns a set of sets of bus ids, each set is a connected component
+"""
+function connected_components(data::Dict{String,Any})
+    if haskey(data, "multinetwork") && data["multinetwork"]
+        error("connected_components does not yet support multinetwork data")
+    end
+
+    active_bus = filter((i, bus) -> bus["bus_type"] != 4, data["bus"])
+    active_bus_ids = Set{Int64}([bus["bus_i"] for (i,bus) in active_bus])
+    #println(active_bus_ids)
+
+    neighbors = Dict(i => [] for i in active_bus_ids)
+    for (i,branch) in data["branch"]
+        if branch["br_status"] != 0 && branch["f_bus"] in active_bus_ids && branch["t_bus"] in active_bus_ids
+            push!(neighbors[branch["f_bus"]], branch["t_bus"])
+            push!(neighbors[branch["t_bus"]], branch["f_bus"])
+        end
+    end
+    for (i,dcline) in data["dcline"]
+        if dcline["br_status"] != 0 && dcline["f_bus"] in active_bus_ids && dcline["t_bus"] in active_bus_ids
+            push!(neighbors[dcline["f_bus"]], dcline["t_bus"])
+            push!(neighbors[dcline["t_bus"]], dcline["f_bus"])
+        end
+    end
+    #println(neighbors)
+
+    component_lookup = Dict(i => Set{Int64}([i]) for i in active_bus_ids)
+    touched = Set{Int64}()
+
+    for i in active_bus_ids
+        if !(i in touched)
+            _dfs(i, neighbors, component_lookup, touched)
+        end
+    end
+
+    ccs = (Set(values(component_lookup)))
+
+    return ccs
+end
+
+
+"""
+performs DFS on a graph
+"""
+function _dfs(i, neighbors, component_lookup, touched)
+    push!(touched, i)
+    for j in neighbors[i]
+        if !(j in touched)
+            new_comp = union(component_lookup[i], component_lookup[j])
+            for k in new_comp
+                component_lookup[k] = new_comp
+            end
+            _dfs(j, neighbors, component_lookup, touched)
         end
     end
 end