+++ test/data.jl
@@ -81,6 +81,139 @@
 end
 
 
+@testset "test topology propagation" begin
+    @testset "component status updates" begin
+        data_initial = PowerModels.parse_file("../test/data/case7_tplgy.m")
+
+        data = PowerModels.parse_file("../test/data/case7_tplgy.m")
+        PowerModels.propagate_topology_status(data)
+
+        @test length(data_initial["bus"]) == length(data["bus"])
+        @test length(data_initial["gen"]) == length(data["gen"])
+        @test length(data_initial["branch"]) == length(data["branch"])
+
+        active_buses = Set(["2", "4", "5", "7"])
+        active_branches = Set(["8"])
+        active_dclines = Set(["3"])
+
+        for (i,bus) in data["bus"]
+            if i in active_buses
+                @test bus["bus_type"] != 4
+            else
+                @test bus["bus_type"] == 4
+            end
+        end
+
+        for (i,branch) in data["branch"]
+            if i in active_branches
+                @test branch["br_status"] == 1
+            else
+                @test branch["br_status"] == 0
+            end
+        end
+
+        for (i,dcline) in data["dcline"]
+            if i in active_dclines
+                @test dcline["br_status"] == 1
+            else
+                @test dcline["br_status"] == 0
+            end
+        end
+    end
+
+    @testset "connecected components" begin
+        data = PowerModels.parse_file("../test/data/case7_tplgy.m")
+        PowerModels.propagate_topology_status(data)
+        cc = PowerModels.connected_components(data)
+
+        cc_ordered = sort(collect(cc); by=length)
+
+        @test length(cc_ordered) == 2
+        @test length(cc_ordered[1]) == 1
+        @test length(cc_ordered[2]) == 3
+
+        active_buses = Set([2, 4, 5, 7])
+
+        for cc in cc_ordered
+            for i in cc
+                @test i in active_buses
+            end
+        end
+    end
+
+    @testset "component filtering updates" begin
+        data_initial = PowerModels.parse_file("../test/data/case7_tplgy.m")
+
+        data = PowerModels.parse_file("../test/data/case7_tplgy.m")
+        PowerModels.propagate_topology_status(data)
+        PowerModels.select_largest_component(data)
+
+        @test length(data_initial["bus"]) == length(data["bus"])
+        @test length(data_initial["gen"]) == length(data["gen"])
+        @test length(data_initial["branch"]) == length(data["branch"])
+
+        active_buses = Set(["4", "5", "7"])
+        active_branches = Set(["8"])
+        active_dclines = Set(["3"])
+
+        for (i,bus) in data["bus"]
+            if i in active_buses
+                @test bus["bus_type"] != 4
+            else
+                @test bus["bus_type"] == 4
+            end
+        end
+
+        for (i,branch) in data["branch"]
+            if i in active_branches
+                @test branch["br_status"] == 1
+            else
+                @test branch["br_status"] == 0
+            end
+        end
+
+        for (i,dcline) in data["dcline"]
+            if i in active_dclines
+                @test dcline["br_status"] == 1
+            else
+                @test dcline["br_status"] == 0
+            end
+        end
+    end
+
+    @testset "output values" begin
+        data = PowerModels.parse_file("../test/data/case7_tplgy.m")
+        PowerModels.propagate_topology_status(data)
+        result = run_opf(data, ACPPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 1778; atol = 1e0)
+
+        solution = result["solution"]
+
+        active_buses = Set(["2", "4", "5", "7"])
+        active_gens = Set(["2", "3"])
+
+        for (i,bus) in data["bus"]
+            if i in active_buses
+                @test !isequal(solution["bus"][i]["va"], NaN)
+            else
+                @test isequal(solution["bus"][i]["va"], NaN)
+            end
+        end
+
+        for (i,gen) in data["gen"]
+            if i in active_gens
+                @test !isequal(solution["gen"][i]["pg"], NaN)
+            else
+                @test isequal(solution["gen"][i]["pg"], NaN)
+            end
+        end
+    end
+
+end
+
+
 @testset "test user ext init" begin
     @testset "3-bus case" begin
         pm = build_generic_model("../test/data/case3.m", ACPPowerModel, PowerModels.post_opf, ext = Dict(:some_data => "bloop"))