+++ test/runtests.jl
@@ -22,11 +22,13 @@ ipopt_solver = IpoptSolver(tol=1e-6, print_level=0)
 pajarito_solver = PajaritoSolver(mip_solver=GLPKSolverMIP(), cont_solver=ipopt_solver, log_level=0)
 scs_solver = SCSSolver(max_iters=1000000, verbose=0)
 
+
+@testset "PowerModels" begin
 include("common.jl")
 
-include("pti.jl")
-
 include("matpower.jl")
+
+include("pti.jl")
 
 include("output.jl")
 
@@ -46,3 +48,4 @@ include("multinetwork.jl")
 
 include("docs.jl")
 
+end